Rename snewell/gcc-9* images to snewell/gcc-alpine
Since alpine doesn't look like it's going to be easy to have multiple
versions of gcc installed, I decided to just use it for whatever it has
in bleeding edge gcc.  I'll pick some other base in the future and use
that to get custom images for different versions of gcc.